[CHORE #160409504]- Move question classes
